Add binary dump from yarvik tab26x stock firmware
This is a dump from the yarvik tab26x livesuit image. It contains the
android 4.0.3 nanda partition and some extra's that where extracted.

This is what is shipped on the device. Yes,
yarvik_tab26x/stock-nanda/linux/linux.bmp is broken and probably not on
purpose, just missed by accident.