Updated to android 4.0.4
This was an official vendor update to android 4.0.4. Some interesting
changes are the addition of several more bitmaps, fixing 'bugs' in
script.fex (sata_used = 0, which makes sense as the tablet does not
have any sata hardware connected). And cleanup of build.prop; They
removed the iNet 97F references and branded it as their own. Also
for some reason it was decided to swap Vol+ and Vol- keys. Highly
annoying.